+++ src/akvq/__main__.py
@@ -1,8 +1,13 @@
 import os
+import subprocess
 from pathlib import Path
 
 from azure.core.exceptions import ClientAuthenticationError
-from azure.identity import DefaultAzureCredential
+from azure.identity import (
+    ChainedTokenCredential,
+    DeviceCodeCredential,
+    InteractiveBrowserCredential,
+)
 from azure.keyvault.secrets import SecretClient
 
 from . import logger
@@ -17,6 +22,7 @@ See :manpage:`bash#name`
 
 akvq-HASS -> HASS (as env)
 hello -> (skip)
+akvq-DEMO-VAR -> DEMO_VAR (as env)
 
 For now I am skipping to implemt auth cache because we are saveing it to disk
 https://github.com/Azure/azure-sdk-for-python/issues/11134
@@ -29,87 +35,105 @@ Uess the same as for ssh keys
 :doc:`azure.keyvault.secrets`
 
 :doc:`azure-keyvault-secrets:index`
-
-:class:`azure.mgmt.resource.subscriptions.SubscriptionClient`
-
-https://docs.microsoft.com/en-us/azure/governance/resource-graph/first-query-python#run-your-first-resource-graph-query
-
-https://stackoverflow.com/q/64654321
-
 """
 
+# start main logger to get all settings
+log = logger.config("akvq")
 
-def main():
+
+def get_secrets() -> dict:
+    """Get all AKVQ secrets from AKV, Expects secret_client"""
+    log.critical("HELLOOOO")
+
+
+def vault_uri() -> str:
+    """
+    Will check in order and return on match:
+
+    - ENV
+    - git config akvq.url
+
+    Returns:
+        str: url of azure key vault
+
+    """
+    uri = os.environ.get("AKVQ_VAULT_URI")
+    if uri:
+        log.debug("Got from ENV")
+
+    if not uri:
+        try:
+            uri = (
+                subprocess.check_output(["git", "config", "akvq.uri"]).decode().strip()
+            )
+            log.debug("Got from git config")
+        except subprocess.CalledProcessError as exit_code:
+            log.debug(exit_code)
+
+    return uri
+
+
+def main() -> int:
     """
     .. todo::
 
         - save secrets to file
         - clear file
         - add config list like poetry
-        - env for vault_url
         - support to add to vault
         - auto find vault based on tag
         - add support for windows
 
-    :class:`azure.identity.DefaultAzureCredential`
-
-    :class:`azure.identity.TokenCachePersistenceOptions`
-
-    `azure-identity logging
-    <https://azuresdkdocs.blob.core.windows.net/$web/python/azure-identity/latest/index.html#logging>`_
-
     """
-    # cache_options = TokenCachePersistenceOptions()
+    akvq_vault_uri = vault_uri()
+    if not akvq_vault_uri:
+        log.critical("Missing vault uri")
+        return 1
 
-    # # From: https://stackoverflow.com/a/65893640
-    # azure_identity_logger = logging.getLogger("azure.identity")
+    interactive_credential = InteractiveBrowserCredential()
+    device_credential = DeviceCodeCredential()
 
-    # # Critical is to hide the error from azure
-    # azure_identity_logger.setLevel(logging.CRITICAL)
+    print(
+        """A web browser has been opened at https://login.microsoftonline.com/organizations/oauth2/v2.0/authorize.
+Please continue the login in the web browser.
 
-    log = logger.config("akvq")
+If no web browser is available or if the web browser fails to open,
+use the device code login."""
+    )
+    # https://azuresdkdocs.blob.core.windows.net/$web/python/azure-identity/latest/index.html#define-a-custom-authentication-flow-with-chainedtokencredential
+    credential_chain = ChainedTokenCredential(interactive_credential, device_credential)
 
-    try:
-        credential = DefaultAzureCredential(
-            # cache_persistence_options=cache_options,
-            exclude_interactive_browser_credential=False,
-            logging_enable=True,
-        )
+    secret_client = SecretClient(
+        vault_url=vault_uri(), credential=credential_chain, logging_enable=True
+    )
 
-        secret_client = SecretClient(
-            vault_url=os.environ.get("AKVQ_VAULT_URI"),
-            credential=credential,
-        )
+    aprefix = "akvq-"
+    secrets = []
+    secret_properties = secret_client.list_properties_of_secrets()
+    for secret_property in secret_properties:
+        # the list doesn't include values or versions of the secrets
+        if secret_property.name.startswith(aprefix):
+            secrets.append(secret_property.name)
 
-        # secret = secret_client.get_secret("hello")
+    akfilename = Path.home() / ".akvq.env"
+    print("Login successful")
 
-        # print(f"secret value from keyvault is: '{secret.value}'")
+    with open(akfilename, "w") as akfile:
+        # Will override oldfile
+        for secret in secrets:
+            real_secret_name = secret.split(aprefix)[1]
 
-        aprefix = "akvq-"
-        secrets = []
-        secret_properties = secret_client.list_properties_of_secrets()
-        for secret_property in secret_properties:
-            # the list doesn't include values or versions of the secrets
-            if secret_property.name.startswith(aprefix):
-                secrets.append(secret_property.name)
+            # AKV names are only allowed to contain alphanumeric characters and dashes
+            # ENV names don't allow dash so a swap works good
+            real_secret_name = real_secret_name.replace("-", "_")
 
-        akfilename = Path.home() / ".akvq.sh"
+            # log.info(f"adding: {real_secret_name}")
+            print(f"adding: {real_secret_name}")
+            secret_value = secret_client.get_secret(secret).value
 
-        with open(akfilename, "w") as akfile:
-            # Will override oldfile
-            for secret in secrets:
-                real_secret_name = secret.split(aprefix)[1]
-                log.info(f"adding: {real_secret_name}")
-                secret_value = secret_client.get_secret(secret).value
-                akfile.write(f"export {real_secret_name}={secret_value}\n")
+            akfile.write(f"export {real_secret_name}={secret_value}\n")
 
-    except ClientAuthenticationError:
-        pass
-    except ValueError as e:
-        log.critical(e)
-
-    # # specify a cache name to isolate the cache from other applications
-    # TokenCachePersistenceOptions(name="akvq")
+    print("You might need to restart your terminal to source the new ENV")
 
 
 if __name__ == "__main__":